+++ tcs-service/src/test/java/com/transformuk/hee/tis/tcs/service/service/impl/PlacementServiceImplTest.java
@@ -700,8 +700,8 @@ public class PlacementServiceImplTest {
   public void isEligibleForChangedWholeTimeEquivalentShouldReturnTrueWhenUpdatedPlacementIsEligibleForNotification() {
     LocalDate dateFiveMonthsAgo = LocalDate.now().minusMonths(5);
     Long existingPlacementId = 1L;
-    BigDecimal existingWholeTimeEquivalent = new BigDecimal(1.0);
-    BigDecimal updatedWholeTimeEquivalent = new BigDecimal(0.5);
+    Float existingWholeTimeEquivalent = new Float(1.0);
+    Float updatedWholeTimeEquivalent = new Float(0.5);
 
     Placement currentPlacement = new Placement();
     currentPlacement.setId(existingPlacementId);
